Change articleid obtaining process, & Add getcachereadarticle

+++ init.js
@@ -16,8 +16,8 @@ async function main() {
     await aquery(`USE ${PICONFIG.DB}`);
     await aquery(`CREATE TABLE Users (Name nvarchar(30) PRIMARY KEY, Hash char(60), Token BINARY(64))`);
     // basically this table only exists to state that each article can only happen once
-    await aquery(`CREATE TABLE Articles (Id smallint unsigned PRIMARY KEY);`);
-    await aquery(`CREATE TABLE UserReadArticles (UserName nvarchar(30), ArticleId smallint unsigned,` + 
+    await aquery(`CREATE TABLE Articles (Id int unsigned PRIMARY KEY);`);
+    await aquery(`CREATE TABLE UserReadArticles (UserName nvarchar(30), ArticleId int unsigned,` + 
         `FOREIGN KEY (UserName) REFERENCES Users(Name),` +
         `FOREIGN KEY (ArticleId) REFERENCES Articles(Id)` +
     `);`);
